Changes to make global run succeed (#37)
* Changes to make global run succeed

- Remove dead code that computes per
  country breaks based on histogram
- Remove tertiary roads at zoom 7, the
  tiles are too big at this zoom and
  you can't see most of these roads
  in large countries
- Put layerToGeoTiff behind debug flag
  because the collect it calls can't
  be run for large countries

* Bump to vectorpipe version using gt 3.1.0

* Add hack to skip countries that cross antimeridian

* Resample to custom extent for RUS and FJI

These countries cross the antimeridian
which throws a exception during image layer
creation.

This workaround allows us to generate tile layers
for most of the extent of these countries
without digging into proj4 reprojection code. The
max 179.7 longitude was arbitrarily selected via
trial and error as a value that doesn't overrun the 180
boundary when the tile layer is buffered later on
in the code.